Update README.md
Here, we proposed a convolutional neural network architecture (here we called RATNet), which includes the concept of the residual unit and attention block with the basic architecture of the convolutional neural network. The model was evaluated for the recognition of handwritten Bengali numerals, modifers, basic and compound characters of three publicly available ISI, CMATERdb, BanglaLekha-Isolated datasets in addition to the proposed dataset. The performance was expressed by three evaluation metrics: Accuracy, F1-score, and Cohen's Kappa.